Move tests from test_metadata_utils.py to test_add_fingerprint.py
